ADR-5/poc/requirements.txt
flask
PyJWT[crypto]
requests